Move SSE and WS clients out of event subscription test
