+++ apps/rig_inbound_gateway/test/event_subscription_test.exs
@@ -7,16 +7,20 @@ defmodule RigInboundGateway.EventSubscriptionTest do
 
   alias HTTPoison
   alias Jason
-  alias Socket.Web, as: WebSocket
 
   alias CloudEvent
   alias RigAuth.Jwt.Utils, as: Jwt
   alias RigInboundGateway.ExtractorConfig
 
-  def external_port,
-    do: Confex.fetch_env!(:rig_inbound_gateway, RigInboundGatewayWeb.Endpoint)[:http][:port]
+  alias SSeClient
+  alias WsClient
 
-  def hostname, do: "localhost"
+  @clients [SseClient, WsClient]
+
+  @hostname "localhost"
+  @eventhub_port Confex.fetch_env!(:rig_inbound_gateway, RigInboundGatewayWeb.Endpoint)[:http][
+                   :port
+                 ]
 
   def setup do
     on_exit(&ExtractorConfig.restore/0)
@@ -38,7 +42,7 @@ defmodule RigInboundGateway.EventSubscriptionTest do
 
   defp update_subscriptions(connection_id, subscriptions, jwt \\ nil) do
     url =
-      "http://#{hostname()}:#{external_port()}/_rig/v1/connection/sse/#{connection_id}/subscriptions"
+      "http://#{@hostname}:#{@eventhub_port}/_rig/v1/connection/sse/#{connection_id}/subscriptions"
 
     body = Jason.encode!(%{"subscriptions" => subscriptions})
 
@@ -51,7 +55,7 @@ defmodule RigInboundGateway.EventSubscriptionTest do
   end
 
   defp submit_event(cloud_event) do
-    url = "http://#{hostname()}:#{external_port()}/_rig/v1/events"
+    url = "http://#{@hostname}:#{@eventhub_port}/_rig/v1/events"
 
     %HTTPoison.Response{status_code: 202} =
       HTTPoison.post!(url, Jason.encode!(cloud_event), [{"content-type", "application/json"}])
@@ -61,153 +65,6 @@ defmodule RigInboundGateway.EventSubscriptionTest do
 
   defp connection_id(welcome_event)
   defp connection_id(%{"data" => %{"connection_token" => connection_id}}), do: connection_id
-
-  defdelegate url_encode_subscriptions(list), to: Jason, as: :encode!
-
-  defmodule Client do
-    @moduledoc false
-    @type client :: pid | reference | atom
-    @callback connect(params :: list) :: {:ok, pid}
-    @callback disconnect(client) :: :ok
-    @callback refute_receive(client) :: :ok
-    @callback read_event(client, event_type :: String.t()) :: map()
-    @callback read_welcome_event(client) :: map()
-    @callback read_subscriptions_set_event(client) :: map()
-  end
-
-  defmodule SseClient do
-    @moduledoc false
-    @behaviour Client
-    alias RigInboundGateway.EventSubscriptionTest, as: Test
-
-    @impl true
-    def connect(params \\ []) do
-      params =
-        if Keyword.has_key?(params, :subscriptions) do
-          encoded_subscriptions = params[:subscriptions] |> Test.url_encode_subscriptions()
-          Keyword.replace!(params, :subscriptions, encoded_subscriptions)
-        else
-          params
-        end
-
-      url =
-        "http://#{Test.hostname()}:#{Test.external_port()}/_rig/v1/connection/sse?#{
-          URI.encode_query(params)
-        }"
-
-      %HTTPoison.AsyncResponse{id: client} =
-        HTTPoison.get!(url, %{},
-          stream_to: self(),
-          recv_timeout: 20_000
-        )
-
-      assert_receive %HTTPoison.AsyncStatus{code: 200}
-      assert_receive %HTTPoison.AsyncHeaders{}
-
-      {:ok, client}
-    end
-
-    @impl true
-    def disconnect(client) do
-      {:ok, ^client} = :hackney.stop_async(client)
-      :ok
-    end
-
-    @impl true
-    def refute_receive(_client) do
-      receive do
-        %HTTPoison.AsyncChunk{} = async_chunk ->
-          raise "Unexpectedly received: #{inspect(async_chunk)}"
-      after
-        100 -> :ok
-      end
-    end
-
-    @impl true
-    def read_event(_client, event_type) do
-      cloud_event = read_sse_chunk() |> extract_cloud_event()
-      %{"eventType" => ^event_type} = cloud_event
-    end
-
-    @impl true
-    def read_welcome_event(client), do: read_event(client, "rig.connection.create")
-
-    @impl true
-    def read_subscriptions_set_event(client), do: read_event(client, "rig.subscriptions_set")
-
-    defp read_sse_chunk do
-      receive do
-        %HTTPoison.AsyncChunk{chunk: chunk} -> chunk
-      after
-        1_000 ->
-          raise "No chunk to read after 1s. #{inspect(:erlang.process_info(self(), :messages))}"
-      end
-    end
-
-    defp extract_cloud_event(sse_chunk) do
-      sse_chunk
-      |> String.split("\n", trim: true)
-      |> Enum.reduce_while(nil, fn
-        "data: " <> data, _acc -> {:halt, Jason.decode!(data)}
-        _x, _acc -> {:cont, nil}
-      end)
-      |> case do
-        nil -> raise "Failed to extract CloudEvent from chunk: #{inspect(sse_chunk)}"
-        cloud_event -> cloud_event
-      end
-    end
-  end
-
-  defmodule WsClient do
-    @moduledoc false
-    @behaviour Client
-    alias RigInboundGateway.EventSubscriptionTest, as: Test
-
-    @impl true
-    def connect(params \\ []) do
-      params =
-        if Keyword.has_key?(params, :subscriptions) do
-          encoded_subscriptions = params[:subscriptions] |> Test.url_encode_subscriptions()
-          Keyword.replace!(params, :subscriptions, encoded_subscriptions)
-        else
-          params
-        end
-
-      WebSocket.connect(Test.hostname(), Test.external_port(), %{
-        path: "/_rig/v1/connection/ws?#{URI.encode_query(params)}",
-        protocol: ["ws"]
-      })
-    end
-
-    @impl true
-    def disconnect(client) do
-      :ok = WebSocket.close(client)
-    end
-
-    @impl true
-    def refute_receive(client) do
-      case WebSocket.recv(client, timeout: 100) do
-        {:ping, _} -> client.refute_receive(client)
-        {:ok, packet} -> raise "Unexpectedly received: #{inspect(packet)}"
-        {:error, _} -> :ok
-      end
-    end
-
-    @impl true
-    def read_event(client, event_type) do
-      {:text, data} = WebSocket.recv!(client)
-      cloud_event = Jason.decode!(data)
-      %{"eventType" => ^event_type} = cloud_event
-    end
-
-    @impl true
-    def read_welcome_event(client), do: read_event(client, "rig.connection.create")
-
-    @impl true
-    def read_subscriptions_set_event(client), do: read_event(client, "rig.subscriptions_set")
-  end
-
-  @clients [SseClient, WsClient]
 
   describe "Connections and subscriptions:" do
     test "Connecting with no parameters causes empty subscription set." do
